build: solve ReactDOM.createRoot warning in dev
The following message was showing up in the console
after changing a file when running the dev server:

```
Warning: You are calling ReactDOMClient.createRoot() on a container that has already been passed to createRoot() before. Instead, call root.render() on the existing root instead if you want to update it.
```